refactor: split oneliner text utils and record store into own files
