+++ internal/menu/executor_oneliners.go
@@ -1,302 +1,18 @@
 package menu
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"io"
 	"log/slog"
-	"os"
 	"path/filepath"
 	"strings"
-	"sync"
 	"time"
-	"unicode/utf8"
 
 	"github.com/ViSiON-3/vision-3-bbs/internal/ansi"
 	"github.com/ViSiON-3/vision-3-bbs/internal/terminalio"
 	"github.com/ViSiON-3/vision-3-bbs/internal/user"
 )
-
-func truncateRunes(value string, max int) string {
-	value = strings.TrimSpace(value)
-	if max <= 0 || value == "" {
-		return ""
-	}
-	if utf8.RuneCountInString(value) <= max {
-		return value
-	}
-	runes := []rune(value)
-	if max <= 3 {
-		return string(runes[:max])
-	}
-	return string(runes[:max-3]) + "..."
-}
-
-func isPipeCodeStartChar(b byte) bool {
-	return (b >= '0' && b <= '9') || (b >= 'A' && b <= 'Z') || (b >= 'a' && b <= 'z')
-}
-
-func pipeCodeLenAt(value string, index int) int {
-	if index < 0 || index >= len(value) || value[index] != '|' || index+1 >= len(value) {
-		return 0
-	}
-
-	// 3-char forms: |00..|15, |CR, |DE, |CL, |PP, |23
-	if index+2 < len(value) {
-		two := value[index+1 : index+3]
-		if len(two) == 2 {
-			if (two[0] >= '0' && two[0] <= '9') && (two[1] >= '0' && two[1] <= '9') {
-				return 3
-			}
-			u := strings.ToUpper(two)
-			if u == "CR" || u == "DE" || u == "CL" || u == "PP" {
-				return 3
-			}
-		}
-	}
-
-	// Background forms: |B0..|B9 (3 chars), |B10..|B15 (4 chars)
-	if index+2 < len(value) {
-		if (value[index+1] == 'B' || value[index+1] == 'b') && (value[index+2] >= '0' && value[index+2] <= '9') {
-			if index+3 < len(value) && (value[index+3] >= '0' && value[index+3] <= '9') {
-				return 4 // |B10..|B15 (validated loosely)
-			}
-			return 3 // |B0..|B9
-		}
-	}
-
-	// 2-char form: |P
-	if index+1 < len(value) && (value[index+1] == 'P' || value[index+1] == 'p') {
-		return 2
-	}
-
-	if index+1 < len(value) && isPipeCodeStartChar(value[index+1]) {
-		return 0
-	}
-
-	return 0
-}
-
-func truncateOnelinerPreservePipeCodes(value string, maxVisible int) string {
-	value = strings.TrimSpace(value)
-	if value == "" || maxVisible <= 0 {
-		return ""
-	}
-
-	var out strings.Builder
-	visible := 0
-	i := 0
-	for i < len(value) {
-		if value[i] == '|' {
-			codeLen := pipeCodeLenAt(value, i)
-			if codeLen > 0 && i+codeLen <= len(value) {
-				out.WriteString(value[i : i+codeLen])
-				i += codeLen
-				continue
-			}
-		}
-
-		r, size := utf8.DecodeRuneInString(value[i:])
-		if r == utf8.RuneError && size == 1 {
-			size = 1
-		}
-		if visible >= maxVisible {
-			break
-		}
-		out.WriteString(value[i : i+size])
-		visible++
-		i += size
-	}
-
-	return strings.TrimSpace(out.String())
-}
-
-func truncatePipeCodedText(value string, maxVisible int) string {
-	if value == "" || maxVisible <= 0 {
-		return ""
-	}
-
-	var out strings.Builder
-	visible := 0
-	i := 0
-	for i < len(value) {
-		if value[i] == '|' && i+1 < len(value) && value[i+1] == '|' {
-			if visible >= maxVisible {
-				break
-			}
-			out.WriteString("||")
-			visible++
-			i += 2
-			continue
-		}
-
-		if value[i] == '|' {
-			codeLen := pipeCodeLenAt(value, i)
-			if codeLen > 0 && i+codeLen <= len(value) {
-				out.WriteString(value[i : i+codeLen])
-				i += codeLen
-				continue
-			}
-		}
-
-		_, size := utf8.DecodeRuneInString(value[i:])
-		if size <= 0 {
-			size = 1
-		}
-		if visible >= maxVisible {
-			break
-		}
-		out.WriteString(value[i : i+size])
-		visible++
-		i += size
-	}
-
-	return out.String()
-}
-
-func containsDisallowedOnelinerColorCode(value string) bool {
-	i := 0
-	for i < len(value) {
-		if value[i] == '|' && i+1 < len(value) && value[i+1] == '|' {
-			i += 2
-			continue
-		}
-
-		if value[i] == '|' {
-			codeLen := pipeCodeLenAt(value, i)
-			if codeLen > 0 && i+codeLen <= len(value) {
-				// Only standard foreground colors |01..|15 are allowed.
-				if codeLen != 3 {
-					return true
-				}
-
-				colorCode := value[i+1 : i+3]
-				if colorCode < "01" || colorCode > "15" {
-					return true
-				}
-
-				i += codeLen
-				continue
-			}
-		}
-
-		_, size := utf8.DecodeRuneInString(value[i:])
-		if size <= 0 {
-			size = 1
-		}
-		i += size
-	}
-
-	return false
-}
-
-func formatOnelinerDisplayName(name string) string {
-	formatted := truncateRunes(name, oneLinerNameWidth)
-	if formatted == "" {
-		formatted = "Unknown"
-	}
-	padding := oneLinerNameWidth - utf8.RuneCountInString(formatted)
-	if padding > 0 {
-		formatted = strings.Repeat(" ", padding) + formatted
-	}
-	return formatted
-}
-
-func onelinerVisibleName(record onelinerRecord, anonymousName string) string {
-	if strings.TrimSpace(anonymousName) == "" {
-		anonymousName = "Anonymous"
-	}
-	if record.Anonymous {
-		return anonymousName
-	}
-	if strings.TrimSpace(record.PostedByHandle) != "" {
-		return record.PostedByHandle
-	}
-	if strings.TrimSpace(record.PostedByUsername) != "" {
-		return record.PostedByUsername
-	}
-	return "Unknown"
-}
-
-func loadOnelinerRecords(onelinerPath string) ([]onelinerRecord, error) {
-	jsonData, readErr := os.ReadFile(onelinerPath)
-	if readErr != nil {
-		if os.IsNotExist(readErr) {
-			return []onelinerRecord{}, nil
-		}
-		return nil, readErr
-	}
-
-	if strings.TrimSpace(string(jsonData)) == "" {
-		return []onelinerRecord{}, nil
-	}
-
-	var rawEntries []json.RawMessage
-	if err := json.Unmarshal(jsonData, &rawEntries); err != nil {
-		return nil, err
-	}
-
-	records := make([]onelinerRecord, 0, len(rawEntries))
-	for _, raw := range rawEntries {
-		var legacyText string
-		if err := json.Unmarshal(raw, &legacyText); err == nil {
-			legacyText = truncateOnelinerPreservePipeCodes(legacyText, oneLinerMaxLength)
-			if legacyText != "" {
-				records = append(records, onelinerRecord{
-					Text:             legacyText,
-					PostedByUsername: "Unknown",
-				})
-			}
-			continue
-		}
-
-		var compat onelinerRecordCompat
-		if err := json.Unmarshal(raw, &compat); err != nil {
-			continue
-		}
-
-		record := onelinerRecord{
-			Text:             truncateOnelinerPreservePipeCodes(compat.Text, oneLinerMaxLength),
-			Anonymous:        compat.Anonymous,
-			PostedByUsername: strings.TrimSpace(compat.PostedByUsername),
-			PostedByHandle:   strings.TrimSpace(compat.PostedByHandle),
-			PostedAt:         compat.PostedAt,
-		}
-
-		if record.PostedByUsername == "" {
-			record.PostedByUsername = strings.TrimSpace(compat.Username)
-		}
-		if record.PostedByHandle == "" {
-			if strings.TrimSpace(compat.DisplayName) != "" && !record.Anonymous {
-				record.PostedByHandle = strings.TrimSpace(compat.DisplayName)
-			} else if strings.TrimSpace(compat.Username) != "" {
-				record.PostedByHandle = strings.TrimSpace(compat.Username)
-			}
-		}
-
-		if record.Text == "" {
-			continue
-		}
-
-		records = append(records, record)
-	}
-
-	return records, nil
-}
-
-func saveOnelinerRecords(onelinerPath string, records []onelinerRecord) error {
-	if len(records) > oneLinerMaxStored {
-		records = records[len(records)-oneLinerMaxStored:]
-	}
-
-	updatedJSON, err := json.MarshalIndent(records, "", "  ")
-	if err != nil {
-		return err
-	}
-
-	return os.WriteFile(onelinerPath, updatedJSON, 0644)
-}
 
 // runOneliners displays the oneliners using templates.
 func runOneliners(c *cmdCtx, args string) (*user.User, string, error) {
@@ -589,32 +305,4 @@ func runOneliners(c *cmdCtx, args string) (*user.User, string, error) {
 	} // end if addYes
 
 	return nil, "", nil
-}
-
-// Mutex for protecting access to the oneliners file
-var onelinerMutex sync.Mutex
-
-const (
-	oneLinerMaxStored  = 20
-	oneLinerMaxDisplay = 10
-	oneLinerMaxLength  = 51
-	oneLinerNameWidth  = 20
-)
-
-type onelinerRecord struct {
-	Text             string `json:"text"`
-	Anonymous        bool   `json:"anonymous,omitempty"`
-	PostedByUsername string `json:"posted_by_username,omitempty"`
-	PostedByHandle   string `json:"posted_by_handle,omitempty"`
-	PostedAt         string `json:"posted_at,omitempty"`
-}
-
-type onelinerRecordCompat struct {
-	DisplayName      string `json:"display_name,omitempty"`
-	Username         string `json:"username,omitempty"`
-	Text             string `json:"text"`
-	Anonymous        bool   `json:"anonymous,omitempty"`
-	PostedByUsername string `json:"posted_by_username,omitempty"`
-	PostedByHandle   string `json:"posted_by_handle,omitempty"`
-	PostedAt         string `json:"posted_at,omitempty"`
 }
